set id to datatype long

+++ src/main/java/com/kinoticket/backend/model/FilmshowSeatPK.java
@@ -4,8 +4,8 @@ import java.io.Serializable;
 
 public class FilmshowSeatPK implements Serializable {
 
-        int seat;
-        int filmShow;
+        long seat;
+        long filmShow;
 
         public boolean equals(Object object) {
             if (object instanceof FilmshowSeatPK) {
@@ -18,8 +18,8 @@ public class FilmshowSeatPK implements Serializable {
 
         public int hashCode() {
             int hash = 23;
-            hash = hash * 31 + seat;
-            hash = hash * 31 + filmShow;
+            hash = hash * 31 + (int) seat;
+            hash = hash * 31 + (int) filmShow;
             return hash;
         }
 }
